+++ plugins/client/grpc/client_config.go
@@ -30,6 +30,7 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/credentials/insecure"
+	_ "google.golang.org/grpc/encoding/gzip" // for install the "gzip" decompressor
 	"google.golang.org/grpc/metadata"
 )