+++ CHANGES.md
@@ -2,13 +2,14 @@ Changes by Version
 ==================
 Release Notes.
 
-1.0.2
+1.1.0
 ------------------
 #### Features
+* Support transmit the OpenTelemetry Metrics protocol.
 
 #### Bug Fixes
 * Fix the missing return data when receive metrics in batch mode.
 
 #### Issues and PR
 - All issues are [here](https://github.com/apache/skywalking/milestone/143?closed=1)
-- All and pull requests are [here](https://github.com/apache/skywalking-satellite/pulls?q=is%3Apr+milestone%3A1.0.2+is%3Aclosed)
+- All and pull requests are [here](https://github.com/apache/skywalking-satellite/pulls?q=is%3Apr+milestone%3A1.1.0+is%3Aclosed)